refactory: rename factory file
